+++ lib/zoanthid/example_group.rb
@@ -2,11 +2,11 @@ module Zoanthid
   module ExampleGroup
     DEFAULT_HOST = "www.example.com"
 
-    def included(base)
-      metadata[:type] = :api
+    def self.included(base)
+      base.metadata[:type] = :api
 
-      if respond_to?(:default_url_options)
-        default_url_options[:host] ||= Zoanthid::ExampleGroup::DEFAULT_HOST
+      if base.respond_to?(:default_url_options)
+        base.default_url_options[:host] ||= Zoanthid::ExampleGroup::DEFAULT_HOST
       end
     end
   end