+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## Unreleased
+
+* Clarify `README`. - Ben Hamill
+* Fix bug in `Zoanthid::ExampleGroup` that caused it to not properly set the
+  type and default host name for specs. - Ben Hamill
+* Over-write the 'request' spec type with 'api' because reasons. - Ben Hamill
+
 ## 0.1.1
 
 * Remove erroneously committed file. - Ben Hamill